- Mise en place du système de requetes de base de données dans un fichier externe au index.js - Recherche des credentials de la base de donnée dans le fichier config.json
